v4/go: fix the structure-table branch span of the validation and recovery walks (SOW-0027 milestone 4)
The dense structure-ID table walkers scaled a directory page's child
bases by the level-0 span (50 IDs) instead of the full coverage of the
level below (50*512^(L-1)); every implied record ID at or above 25,600
was misattributed and valid three-level tables validated with
ReasonStructureInvalid findings. The writer and the reader already used
the correct span; only the validation walk, the recovery scan, and one
test descender had the off-by-one against the Rust authority
(structured_value/table.rs coverage(level-1)).

Add level-two-root regression tests for the validation walk and the
recovery structure count (synthetic generations with records at ids 1
and 25,600; both fail on the old span and pass on the fixed span), and
repair the bench harness's seed handling for counts above 2^63, which
previously produced a false feed-first-overlap writer alarm. The Go
scale matrix now passes structured-build-random up to 1M records;
matched full-matrix evidence is recorded in the SOW status.

+++ v4/go/cmd/iprange-v4-bench/source.go
@@ -24,7 +24,10 @@ func newPermutation(count int, seed uint64) permutation {
 	if count <= 1 {
 		return permutation{count: count, step: 1, offset: 0}
 	}
-	step := int(seed) % count
+	// Rust uses usize arithmetic on the u64 seed (identity cast), so the
+	// modulo runs in uint64 before the conversion to int; a direct
+	// int(seed) would go negative for seeds >= 2^63.
+	step := int(seed % uint64(count))
 	if step == 0 {
 		step = 1
 	}
@@ -37,7 +40,7 @@ func newPermutation(count int, seed uint64) permutation {
 	return permutation{
 		count:  count,
 		step:   step,
-		offset: int(rotl64(seed, 17)) % count,
+		offset: int(rotl64(seed, 17) % uint64(count)),
 	}
 }
